% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-Constant.R
\docType{class}
\name{Constant-class}
\alias{Constant}
\alias{Constant-class}
\title{Constant scalars and vectors}
\description{
This is a Reference Class for fixed scalars and vectors
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}
\section{Methods}{

\describe{
\item{\code{get.extension(path)}}{Provides the extension of a file}

\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE, conf = NULL)}}{Default method to initialize basic objects}
}}

