% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-Derived.R
\docType{class}
\name{Derived-class}
\alias{Derived}
\alias{Derived-class}
\title{Derived scalars and vectors}
\description{
This is a Reference Class for derived model quantities
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}
\section{Methods}{

\describe{
\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE)}}{Default method to initialize basic objects}
}}

