% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-LogLikelihood.R
\docType{class}
\name{LogLikelihood-class}
\alias{LogLikelihood}
\alias{LogLikelihood-class}
\title{Model log-likelihood}
\description{
This is a Reference Class for the log-likelihood
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}
\section{Methods}{

\describe{
\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE)}}{Default method to initialize basic objects}
}}

