% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-ModelParameter.R
\docType{class}
\name{ModelParameter-class}
\alias{ModelParameter}
\alias{ModelParameter-class}
\title{Model parameter}
\description{
This is a Reference Class for random variables and vectors that enter models
}

