% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-Normal.R
\docType{class}
\name{Normal-class}
\alias{Normal}
\alias{Normal-class}
\title{Normal random variables and vectors}
\description{
This is a Reference Class for normal random variables and vectors
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}
\section{Methods}{

\describe{
\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE, conf = NULL)}}{Default method to initialize basic objects}
}}

