% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-basic.R
\docType{class}
\name{rcvirtual.basic-class}
\alias{rcvirtual.basic-class}
\title{Basic virtual reference class}
\description{
Virtual ("template") Reference Class for all RCs
}
\details{
This reference class contains fields (aka "attributes")
and methods (aka "procedures") for all basic RCs.
}
\section{Fields}{

\describe{
\item{\code{package.name}}{character. Name of package}

\item{\code{object.name}}{character. Name of object}

\item{\code{verbose}}{logical. Are methods verbose when called?

#@import ncdf4}
}}
\section{Methods}{

\describe{
\item{\code{construct()}}{Construct basic objects}

\item{\code{fields()}}{Lists the fields available in this object}

\item{\code{get.args(function.name.pattern)}}{Lists the arguments in all the exclusive functions
that match the pattern provided}

\item{\code{get.csv(fullpath, unlist = FALSE)}}{General method to import comma separated value files}

\item{\code{get.name()}}{Provides the object's name.}

\item{\code{get.rdata(fullpath)}}{Reads an .RData or .rda file and passes
its contents as a list}

\item{\code{get.txt(fullpath)}}{Reads a .txt file and passes
its contents as a dataframe}

\item{\code{help(method = .self$methods())}}{Prints the description under a specific method}

\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE)}}{Default method to initialize basic objects}

\item{\code{methods()}}{Lists the methods available in this object}

\item{\code{set.name(new.name)}}{Sets the name of a random variable/vector.}

\item{\code{set.verbose(verbose)}}{Changes verbosity level}

\item{\code{validate()}}{Validate basic objects}
}}

