% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factory.R
\docType{class}
\name{rcvirtual.factory-class}
\alias{rcvirtual.factory-class}
\title{Constructor (factory) of statistical models}
\description{
This is a virtual Reference Class for concrete factory RCs
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any factory RC must have.
}

