% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-ggplotter.R
\docType{class}
\name{rcvirtual.ggplotter-class}
\alias{rcvirtual.ggplotter-class}
\title{rcvirtual.ggplotter}
\description{
Reference Class for plotting data with ggplot
}
\details{
This reference class contains fields (aka "attributes")
and methods (aka "procedures") specific to generating
ggplots of data.
}
\section{Methods}{

\describe{
\item{\code{density.plot(out, mytitle, dolegend = TRUE, xpos = 1, ypos = 1)}}{Plots the posterior density of >= 1 model parameters}

\item{\code{lineplot(specs, xpos = 1, ypos = 1)}}{Multiple line plot}

\item{\code{one.density.plot(out, mytitle, xpos = 1, ypos = 1)}}{Plots the posterior density of one hyperparameter}

\item{\code{probability.1dplot(specs, mypoints, xpos = 1, ypos = 1)}}{Plot of the posterior/forecast distribution at a
single site, over time}

\item{\code{qqplot(specs, xpos = 1, ypos = 1)}}{Quantile-quantile plot}

\item{\code{scatterplot(out, mytitle, xpos = 1, ypos = 1)}}{Plots a scatterplot of two model parameters}

\item{\code{sectionplot(specs, xpos = 1, ypos = 1, mypoints = NULL)}}{Section plotter for the vectors specs$x or specs$y
(one of the 2 is a scalar) and specs$z.
The argument specs$labels is a character vector
(size 4) with title and axes labels.}

\item{\code{surfaceplot(specs, xpos = 1, ypos = 1, mypoints = NULL)}}{Surface plotter for the vectors specs$x, specs$y and
specs$z. The argument specs$labels is a character vector
(size 4) with title and axes labels.}

\item{\code{surfaceplot2(base, under, main, top, xpos = 1, ypos = 1)}}{Surface plotter, version 2}

\item{\code{tsplot(out, mytitle, mylabs, xpos = 1, ypos = 1)}}{Plots a time series plot with observed and fit
(median plus 95 CI)}
}}

