% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-gui.R
\docType{class}
\name{rcvirtual.gui-class}
\alias{rcvirtual.gui-class}
\title{Reference class controller of a Shiny app}
\description{
Reference class controller of a Shiny app
}
\section{Fields}{

\describe{
\item{\code{df}}{data.frame.}

\item{\code{previous.txt}}{character. Previously edited value.}
}}
\section{Methods}{

\describe{
\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE, conf = NULL)}}{Default method to initialize basic objects}
}}

