% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-netcdf.R
\docType{class}
\name{rcvirtual.netcdf-class}
\alias{rcvirtual.netcdf-class}
\title{rcvirtual.netcdf: virtual RC that handles netcdf files}
\description{
rcvirtual.netcdf: virtual RC that handles netcdf files
}
\section{Fields}{

\describe{
\item{\code{location}}{character.}

\item{\code{mydata}}{ANY.}

\item{\code{is.spatial.avg}}{logical.}

\item{\code{is.time.indexed}}{logical.}

\item{\code{dim.names}}{character.}

\item{\code{var.names}}{character.}

\item{\code{var.units}}{character.}

\item{\code{var.descr}}{character.}

\item{\code{lon.name}}{character.}

\item{\code{lat.name}}{character.}

\item{\code{time.name}}{character.}

\item{\code{lon.bounds}}{numeric.}

\item{\code{lat.bounds}}{numeric.}

\item{\code{spatial.resolution}}{numeric.}

\item{\code{time.bounds}}{POSIXlt.}

\item{\code{lon360}}{logical.}

\item{\code{start.date}}{POSIXlt.}

\item{\code{time.units}}{character.}

\item{\code{time.mult}}{numeric.}

\item{\code{n.spatial}}{numeric.}

\item{\code{instants}}{POSIXlt.}
}}
\section{Methods}{

\describe{
\item{\code{construct()}}{Construct basic objects}

\item{\code{get.dim.values(dim.name)}}{Array of values of a specified dimension}

\item{\code{get.mydata()}}{Returns data from object}

\item{\code{get.name()}}{Provides the object's name.}

\item{\code{get.summary()}}{URL, variable names, units, descriptions.}

\item{\code{get.var.names()}}{List of variable names}

\item{\code{get.xyz(x.name = "lon", y.name = "lat", z.name, include.z = NULL,
  subsample.treshold = 1000, bounds = NULL)}}{Provides a 3-column data frame (x,y,z) from a netcdf
file, e.g. for plotting. Subsamples if needed.}

\item{\code{iget.2Dslice.grid(var.name = NULL, xbounds360 = NULL, ybounds = NULL)}}{Get slice of data from 2D variable (lat/lon)}

\item{\code{iget.2Dslice.station(var.name = NULL, xbounds360 = NULL, ybounds = NULL,
  tidx = NULL)}}{Get a slice of data from a specified 2D variable
(time and location index)}

\item{\code{iget.3Dslice(var.name = NULL, xbounds360 = NULL, ybounds = NULL,
  tidx = NULL)}}{Get slice of data from a 3D variable (time/lat/lon)}

\item{\code{iget.spatial.bounds(xbounds360, ybounds)}}{Sets up spatial bounds for slicing}

\item{\code{iget.time.indices(tbounds)}}{Sets up time bound indices for slicing}

\item{\code{set.lon360(value)}}{Allows users to override the value of lon360,
which defines whether longitude ranges betweeen
-180 to 180 (FALSE) or 0 to 360 (TRUE),
in case the automated algorithm gets it wrong.}

\item{\code{set.slice(specs = NULL)}}{Takes (a slice from) the input netcdf file associated
with this object and writes it to a new file.
If specs are not provided, this method copy-pastes
the 1st variable to a temporary folder}
}}

