% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameters.R
\docType{class}
\name{rcvirtual.parameters-class}
\alias{rcvirtual.parameters-class}
\title{Constants, random variables and vectors (parameters) of statistical models}
\description{
This is a virtual Reference Class for concrete parameter RCs
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any parameter RC must have.
}

