% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotter.R
\docType{class}
\name{rcvirtual.plotter-class}
\alias{rcvirtual.plotter-class}
\title{Graphics controller (plotter) of statistical models}
\description{
This is a virtual Reference Class for concrete plotter RCs
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any plotter RC must have.
}
\section{Methods}{

\describe{
\item{\code{construct(specs)}}{Construct basic objects}

\item{\code{get.buffer.plot()}}{Prints the plots in the buffer}

\item{\code{set.buffer.size(nr, nc)}}{Sets up plotter's buffer size (number of rows and columns)}

\item{\code{set.in.buffer(myplot, xpos, ypos)}}{Places a plot in the buffer}

\item{\code{set.palette(argument, value)}}{Sets the printer's colour palettes}
}}

