% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-random.R
\docType{class}
\name{rcvirtual.random-class}
\alias{rcvirtual.random-class}
\title{Random variables and vectors}
\description{
This is a virtual Reference Class for random variable and random vector RCs
}
\section{Methods}{

\describe{
\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE)}}{Default method to initialize basic objects}

\item{\code{reset.rng()}}{Resets this object's random number generator}

\item{\code{size()}}{Provides the size of this object.}
}}

