% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcvirtual.R
\docType{package}
\name{rcvirtual}
\alias{rcvirtual}
\alias{rcvirtual-package}
\title{rcvirtuak: foundations for statistical models with R reference classes}
\description{
This package establishes the foundations of several reference classes that
can be employed to set up, fit and explore a statistical model
}
\details{
The following virtual reference classes are included in this package:
1) the daemon is the RC that has overall control over the model, and is the
entity that users should interact with to get things done;
2) the factory RC creates appropriate strategy and parameter objects and
passes them to the daemon;
3) parameter RCs contain all the information about model parameters;
4) strategy RCs define how the model is fitted
5) plotter RCs specify plotting functions
}

