% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-Time.R
\docType{class}
\name{Time-class}
\alias{Time}
\alias{Time-class}
\title{Time objects}
\description{
This is a Reference Class for objects that handle time information
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}
\section{Methods}{

\describe{
\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE)}}{Default method to initialize basic objects}
}}

