% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-TruncatedNormal.R
\docType{class}
\name{TruncatedNormal-class}
\alias{TruncatedNormal}
\alias{TruncatedNormal-class}
\title{Truncated Normal random variables and vectors}
\description{
This is a Reference Class for truncated normal random variables and vectors
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}

