% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-Uniform.R
\docType{class}
\name{Uniform-class}
\alias{Uniform}
\alias{Uniform-class}
\title{Uniform random variables}
\description{
This is a Reference Class for scalar Uniform random variables
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}

