% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-GaussianState.R
\docType{class}
\name{GaussianState-class}
\alias{GaussianState}
\alias{GaussianState-class}
\title{Gaussian State variables and vectors}
\description{
This is a Reference Class for Dynamic (Non-)Linear Models
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}

