% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-InverseGamma.R
\docType{class}
\name{InverseGamma-class}
\alias{InverseGamma}
\alias{InverseGamma-class}
\title{Inverse Gamma random variables}
\description{
This is a Reference Class for scalar inverse Gamma random variables
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any random RC must have.
}

