% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d-ModelParameter.R
\docType{class}
\name{ModelParameter-class}
\alias{ModelParameter}
\alias{ModelParameter-class}
\title{Model parameter}
\description{
This is a Reference Class for random variables and vectors that enter models
}
\section{Methods}{

\describe{
\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE)}}{Default method to initialize basic objects}
}}

