% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic.R
\docType{class}
\name{rcvirtual.basic-class}
\alias{rcvirtual.basic-class}
\title{Basic virtual reference class}
\description{
Virtual ("template") Reference Class for all reference classes
}
\details{
This reference class contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for all basic RCs.
}
\section{Fields}{

\describe{
\item{\code{name}}{character. Name of object}

\item{\code{verbose}}{logical. Should methods be verbose when they run?}
}}
\section{Methods}{

\describe{
\item{\code{activate(i)}}{Lists available methods, other than standard ones}

\item{\code{construct(specs)}}{Construct basic objects}

\item{\code{fields()}}{Lists the fields available in this object}

\item{\code{help(method = .self$methods())}}{Prints the description under a specific method}

\item{\code{initialize(name = "anonymous", verbose = TRUE, autoconstruct = FALSE,
  specs = NULL)}}{Default method to initialize basic objects}

\item{\code{methods()}}{Lists the methods available in this object}

\item{\code{set.verbose(verbose)}}{Changes verbosity level}
}}

