% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-conf.R
\docType{class}
\name{rcvirtual.conf-class}
\alias{rcvirtual.conf-class}
\title{rcvirtual.conf: Object configuration}
\description{
rcvirtual.conf: Object configuration
}
\section{Fields}{

\describe{
\item{\code{daemon.}}{Specifications for daemon}

\item{\code{parameters.}}{Specifications for parameters}

\item{\code{strategy.}}{Specifications for strategy}

\item{\code{plotter.}}{Specifications for plotter}
}}
\section{Methods}{

\describe{
\item{\code{construct()}}{Construct basic objects}
}}

