% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/daemon.R
\docType{class}
\name{rcvirtual.daemon-class}
\alias{rcvirtual.daemon-class}
\title{Controller (daemon) of statistical models}
\description{
This is a virtual Reference Class for concrete daemon RCs
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any daemon RC must have.
}

