% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-daemon.R
\docType{class}
\name{rcvirtual.daemon-class}
\alias{rcvirtual.daemon-class}
\title{Controller (daemon) of statistical models}
\description{
This is a virtual Reference Class for concrete daemon RCs
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any daemon RC must have.
}
\section{Fields}{

\describe{
\item{\code{parameters.}}{Parameters object}

\item{\code{strategy.}}{Strategy object}

\item{\code{plotter.}}{Plotter object}
}}
\section{Methods}{

\describe{
\item{\code{clean.tmp()}}{Clean the temporary folder}

\item{\code{construct()}}{Construct basic objects}

\item{\code{fit(mle = TRUE)}}{Request the daemon to fit a model}

\item{\code{get.data(name)}}{Retrieve data, stored in memory and controlled
by the daemon}

\item{\code{initialize(package.name = "anonymous", object.name = "anonymous",
  verbose = TRUE, autoconstruct = FALSE, conf = NULL)}}{Default method to initialize basic objects}

\item{\code{vignette(package.name = .self$name)}}{Provides the vignette associated with this package}
}}

