% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-parameters.R
\docType{class}
\name{rcvirtual.parameters-class}
\alias{rcvirtual.parameters-class}
\title{Constants, random variables and vectors (parameters) of
statistical models
This is a virtual Reference Class for parameter RCs}
\description{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any parameter RC must have
}
\section{Fields}{

\describe{
\item{\code{name}}{character.}

\item{\code{netcdf.io}}{list.}

\item{\code{value}}{list.}

\item{\code{count}}{list.}

\item{\code{mean}}{list.}

\item{\code{variance}}{list.}

\item{\code{diff}}{list.}
}}
\section{Methods}{

\describe{
\item{\code{construct()}}{Construct basic objects}

\item{\code{get.data(param.name = NULL, long.name = NULL, field.name = "value")}}{param.name = name of parameter (alpha, A, etc)
field.name = name of parameter's property
(prior mean, lbound, value, etc.)}

\item{\code{get.density(variate, distr, mean, var, lb, ub, dolog)}}{Computes a p.d.f., based on input specifications}

\item{\code{get.id(param.name = NULL, long.name = NULL)}}{Given a name, provides the parameter ID.
Works for vectors of names}

\item{\code{get.lat.bounds()}}{Provides the min/max latitude of data used in model}

\item{\code{get.log.prior.density(param.name, variate)}}{Computes the log-prior density}

\item{\code{get.log.proposal.density(param.name, variate)}}{Computes the log-proposal density}

\item{\code{get.lon.bounds()}}{Provides the min/max longitude of data used in model}

\item{\code{get.names(id = NULL, long = FALSE)}}{Provides the name of the object with index id}

\item{\code{get.size(id = NULL, param.name = NULL, long.name = NULL)}}{Provides the size of an object, identified by one of
the following: id, name or long.name}

\item{\code{get.time(idx = NULL, frm = "instants")}}{Retrieves time information from object}

\item{\code{get.time.bounds(fmt = "POSIXlt")}}{Lower and upper time bounds
(days since Jan 01, 1970)}

\item{\code{get.time.step(frm = "character")}}{Provides the time step of the model, either as a
string ('year', 'month', 'day'), or as a
number [days] (1, 31, 365).}

\item{\code{get.unk.parameters()}}{Provides a list of names of parameters whose
MLE/MAP/posterior will be or has been computed,
as well as their current values and
upper/lower bounds}

\item{\code{get.value(param.name = NULL, long.name = NULL)}}{Rapid access to parameter value}

\item{\code{is.redundant.calculation(param.name = NA, parent.names = NULL)}}{Should a derived quantity be recomputed?}

\item{\code{is.valid()}}{Function that checks if parameters are valid.}

\item{\code{is.valid.field(field.name)}}{Tests whether fname is among the field names of
this object}

\item{\code{is.valid.parameter(param.name)}}{Tests whether fname is among the parameter names of
this object}

\item{\code{set.lat.bounds(lat.bounds)}}{Sets the min/max latitude of data used in model}

\item{\code{set.lon.bounds(lon.bounds)}}{Sets the min/max latitude of data used in model}

\item{\code{set.read.data()}}{Reads data from files}

\item{\code{set.time()}}{Function that stores time information into
     rcvirtual.parameters object: 
instants,
     days, months, years, dates, starting instant and
     ending instant.}

\item{\code{set.value(param.name = NULL, long.name = NULL, objs)}}{Shortcut method to set parameter/config values}
}}

