% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strategy.R
\docType{class}
\name{rcvirtual.strategy-class}
\alias{rcvirtual.strategy-class}
\title{Statistical model-fitting algorithms (strategy)}
\description{
This is a virtual Reference Class for concrete strategy RCs
}
\details{
This RC contains fields (a.k.a. "attributes") and methods
(a.k.a. "procedures") for that any strategy RC must have.
}
\section{Methods}{

\describe{
\item{\code{construct(specs)}}{Construct basic objects}

\item{\code{is.model.fitted()}}{Check if the model has been fitted}

\item{\code{set.model.fitted(status)}}{Change the status of the model (fitted vs. not fitted)}

\item{\code{validate()}}{Function that checks if strategy is valid.}
}}

